+++ src/Symfony/Bridge/Doctrine/ManagerRegistry.php
@@ -16,6 +16,7 @@ use ProxyManager\Proxy\GhostObjectInterface;
 use ProxyManager\Proxy\LazyLoadingInterface;
 use Symfony\Bridge\ProxyManager\LazyProxy\Instantiator\RuntimeInstantiator;
 use Symfony\Component\DependencyInjection\Container;
+use Symfony\Component\VarExporter\LazyGhostObjectInterface;
 
 /**
  * References Doctrine connections and entity/document managers.
@@ -47,8 +48,15 @@ abstract class ManagerRegistry extends AbstractManagerRegistry
         }
         $manager = $this->container->get($name);
 
+        if ($manager instanceof LazyGhostObjectInterface) {
+            if (!$manager->resetLazyGhostObject()) {
+                throw new \LogicException(sprintf('Resetting a non-lazy manager service is not supported. Declare the "%s" service as lazy.', $name));
+            }
+
+            return;
+        }
         if (!$manager instanceof LazyLoadingInterface) {
-            throw new \LogicException('Resetting a non-lazy manager service is not supported. '.(interface_exists(LazyLoadingInterface::class) && class_exists(RuntimeInstantiator::class) ? sprintf('Declare the "%s" service as lazy.', $name) : 'Try running "composer require symfony/proxy-manager-bridge".'));
+            throw new \LogicException(sprintf('Resetting a non-lazy manager service is not supported. Declare the "%s" service as lazy.', $name));
         }
         if ($manager instanceof GhostObjectInterface) {
             throw new \LogicException('Resetting a lazy-ghost-object manager service is not supported.');