+++ src/Symfony/Bridge/ProxyManager/Tests/LazyProxy/Dumper/PhpDumperTest.php
@@ -13,7 +13,6 @@ namespace Symfony\Bridge\ProxyManager\Tests\LazyProxy\Dumper;
 
 use PHPUnit\Framework\TestCase;
 use ProxyManager\Proxy\LazyLoadingInterface;
-use Symfony\Bridge\ProxyManager\LazyProxy\PhpDumper\ProxyDumper;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Dumper\PhpDumper;
 
@@ -62,13 +61,11 @@ class PhpDumperTest extends TestCase
     {
         $container = new ContainerBuilder();
 
-        $container->register('foo', 'stdClass')->setPublic(true);
-        $container->getDefinition('foo')->setLazy(true);
+        $container->register('foo', \stdClass::class)->setPublic(true);
+        $container->getDefinition('foo')->setLazy(true)->addTag('proxy', ['interface' => \stdClass::class]);
         $container->compile();
 
         $dumper = new PhpDumper($container);
-
-        $dumper->setProxyDumper(new ProxyDumper());
 
         return $dumper->dump(['class' => 'LazyServiceProjectServiceContainer']);
     }